Refs #31445. Move fitting contexts to separate folder.
